+++ R/ppc-errors.R
@@ -19,6 +19,7 @@
 #'   [ggplot2::geom_point()] to control the appearance of the points. For the
 #'   binned error plot, arguments controlling the size of the outline and
 #'   opacity of the shaded region indicating the 2-SE bounds.
+#' @param linewidth For the binned error plot, the line width of the outline.
 #'
 #' @details
 #' All of these functions (aside from the `*_scatter_avg` functions)
@@ -341,9 +342,11 @@ ppc_error_binned <-
            ...,
            facet_args = list(),
            bins = NULL,
-           size = 1,
+           size = NULL,
+           linewidth = 1,
            alpha = 0.25) {
     check_ignored_arguments(...)
+    linewidth <- resolve_linewidth(size, linewidth, default_linewidth = 1, calling_fn = "ppc_error_binned")
 
     qx <- enquo(x)
     data <- ppc_error_binnned_data(y, yrep, x = x, bins = bins)
@@ -369,12 +372,12 @@ ppc_error_binned <-
       geom_path(
         mapping = aes(y = .data$se2),
         color = get_color("l"),
-        linewidth = size
+        linewidth = linewidth
       ) +
       geom_path(
         mapping = aes(y = -.data$se2),
         color = get_color("l"),
-        linewidth = size
+        linewidth = linewidth
       ) +
       geom_point(
         mapping = aes(y = .data$err_bar),